+++ jdbc/src/SqlExpr.kt
@@ -50,9 +50,11 @@ infix fun <T, V> KProperty1<T, V>.lt(value: V) = this to SqlOp("<", value)
 infix fun <T, V> KProperty1<T, V>.lte(value: V) = this to SqlOp("<=", value)
 infix fun <T, V> KProperty1<T, V>.like(value: String) = this to SqlOp("like", value)
 infix fun <T, V> KProperty1<T, V>.ilike(value: String) = this to SqlOp("ilike", value)
-infix fun <T, V> KProperty1<T, V>.any(value: Any) = this to SqlExpr("?=any($name)", value)
+infix fun <T, V> KProperty1<T, V>.any(value: V) = this to SqlExpr("?=any($name)", value)
 infix fun <T, V: Comparable<V>> KProperty1<T, V>.between(range: ClosedRange<V>) = this to Between(range.start, range.endInclusive)
 infix fun <T, V: Comparable<V>> KProperty1<T, V>.between(range: OpenEndRange<V>) = this to BetweenExcl(range.start, range.endExclusive)
+infix fun <T, E, V: Collection<E>> KProperty1<T, V>.`in`(values: V) = this to In(values)
+infix fun <T, E, V: Collection<E>> KProperty1<T, V>.notIn(values: V) = this to NotIn(values)
 
 class Between(from: Comparable<*>, to: Comparable<*>): SqlExpr("", from, to) {
   constructor(range: ClosedRange<*>): this(range.start, range.endInclusive)