+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-* jdbc: between operator introduced with open and closed ranges
+* jdbc: between operator introduced with open and closed ranges, also in and notIn
 
 # 1.6.9
 * core: logger(name) is now accessible without a class instance context