+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * jdbc: fixed usage of multiple different DataSources when there is an active transaction
 * json: TSGenerator will now use more type-safe string template types for java.time classes, e.g. `${number}-${number}-${number}` instead of `string`
 * server: `Server(InetSocketAddress(0))` can now be used to bind to any available port.
+* server: make development and test restarts quicker by not using graceful delay #96
 * jackson: serialize enums using their toString() method by default, this fixes `openapi` module usage with `jackson` #88
 * liquibase: do not close jdbc connection if it was passed by user #81